requirements.txt
selenium~=4.8.0
pytest~=7.2.1
webdriver-manager~=3.8.5
pytest-rerunfailures~=11.1.2
pytest-html-reporter~=0.2.9
allure-pytest